requirements.txt
ansi2html==1.8.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.4.1
docstring-parser==0.15
et-xmlfile==1.1.0
eventlet==0.33.3
Flask==2.2.5
Flask-SocketIO==5.1.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.25.1
openpyxl==3.1.5
packaging==23.1
pandas==2.0.3
paramiko==3.2.0
plotly==5.15.0
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
rich==13.7.0
shtab==1.7.0
six==1.16.0
tenacity==8.2.2
typing_extensions==4.7.1
tyro==0.7.3
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
